Warning cleanup
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23711 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/Streamline/avtStreamlineRenderer.C
@@ -261,7 +261,7 @@ avtStreamlineRenderer::SetAtts(const AttributeGroup *a)
 // ****************************************************************************
 
 void
-avtStreamlineRenderer::GlobalSetAmbientCoefficient(float coeff)
+avtStreamlineRenderer::GlobalSetAmbientCoefficient(double coeff)
 {
     ambient_coeff = coeff;
 }
